Some fixes on stack and cpu_proc

+++ code/include/stack.h
@@ -3,7 +3,7 @@
 #include <common.h>
 
 void stack_push(u8 data);
-void stack_push16(u8 data);
+void stack_push16(u16 data);
 
 u8 stack_pop();
 u16 stack_pop16();
